Implementing permission based access

+++ app/__init__.py
@@ -6,12 +6,12 @@ from config import Config
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
 from flask_login import LoginManager
-from flask_security import RoleMixin, UserMixin, Security, SQLAlchemyUserDatastore
-from flask_security.utils import encrypt_password
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_admin import Admin, helpers as admin_helpers
 from flask_mail import Mail
 from flask_babel import Babel, lazy_gettext as _l
+from flask_admin.contrib.sqla import ModelView
+
 
 
 db = SQLAlchemy()
@@ -20,12 +20,8 @@ login = LoginManager()
 login.login_view = 'auth.login'
 login.login_message = 'Please log in to access this page.'
 babel = Babel()
-admin = Admin()
+_admin = Admin()
 mail = Mail()
-
-from app.models import User, Role
-user_datastore = SQLAlchemyUserDatastore(db, User, Role)
-security = Security()
 
 
 def create_app(Config_class=Config):
@@ -36,11 +32,13 @@ def create_app(Config_class=Config):
     migrate.init_app(app, db)
     login.init_app(app)
     babel.init_app(app)
-    security_state = security.init_app(app, user_datastore)
-    security._state = security_state
     
+    from app.models import User, Role
+    from app.admin.views import UserModelView, _AdminIndexView
 
-    admin.init_app(app)
+    _admin.init_app(app, index_view=_AdminIndexView())
+    _admin.add_view(UserModelView(User, db.session))
+    _admin.add_view(UserModelView(Role, db.session))
     mail.init_app(app)
 
     from app.errors import bp as errors_bp
@@ -84,15 +82,7 @@ def create_app(Config_class=Config):
 
     return app
 
-
-@security.context_processor
-def security_context_processor():
-    return dict(
-        admin_base = admin.base_template,
-        admin_view = admin.index_view,
-        h = admin_helpers,
-        get_url = url_for
-    )
+from app.models import User, Role
 
 app = create_app()
 
@@ -101,50 +91,56 @@ def build_sample_db():
     """
     Populate a small db with some example entries.
     """
+    user = db.session.query(User).first()
+    if user is None:
+        import string
+        import random
 
-    import string
-    import random
+        db.drop_all()
+        db.create_all()
 
-    db.drop_all()
-    db.create_all()
+        with current_app.app_context():
+            user_role = Role(name='user', description='Normal user')
+            super_user_role = Role(name='superuser', description='Super user')
+            db.session.add(user_role)
+            db.session.add(super_user_role)
+            db.session.commit()
 
-    with current_app.app_context():
-        user_role = Role(name='user')
-        super_user_role = Role(name='superuser')
-        db.session.add(user_role)
-        db.session.add(super_user_role)
-        db.session.commit()
-
-        text_user = user_datastore.create_user(
-            first_name='Admin',
-            email='admin@admin.com',
-            password_hash=generate_password_hash('admin'),
-            roles=[user_role, super_user_role]
-        )
-
-        first_names = [
-            'Harry', 'Amelia', 'Oliver', 'Jack', 'Isabella', 'Charlie', 'Sophie', 'Mia',
-            'Jacob', 'Thomas', 'Emily', 'Lily', 'Ava', 'Isla', 'Alfie', 'Olivia', 'Jessica',
-            'Riley', 'William', 'James', 'Geoffrey', 'Lisa', 'Benjamin', 'Stacey', 'Lucy'
-        ]
-
-        last_names = [
-            'Brown', 'Smith', 'Patel', 'Jones', 'Williams', 'Johnson', 'Taylor', 'Thomas',
-            'Roberts', 'Khan', 'Lewis', 'Jackson', 'Clarke', 'James', 'Phillips', 'Wilson',
-            'Ali', 'Mason', 'Mitchell', 'Rose', 'Davis', 'Davies', 'Rodriguez', 'Cox', 'Alexander'
-        ]
-
-        for i in range(len(first_names)):
-            tmp_email = first_names[i].lower() + "." + last_names[i].lower() + "@example.com"
-            tmp_pass = ''.join(random.choice(string.ascii_lowercase + string.digits) for i in range(10))
-            user_datastore.create_user(
-                first_name = first_names[i],
-                last_name = last_names[i],
-                email = tmp_email,
-                password_hash = generate_password_hash(tmp_pass),
-                roles = [user_role, ]
+            super_user = User(
+                first_name='Arnold',
+                last_name='Muriuki',
+                middle_name='Nderitu',
+                username='amuriuki',
+                email='arnoldnderitu@gmail.com',
+                password_hash=generate_password_hash('admin'),
+                roles=[user_role, super_user_role]
             )
-        db.session.commit()
+
+            first_names = [
+                'Harry', 'Amelia', 'Oliver', 'Jack', 'Isabella', 'Charlie', 'Sophie', 'Mia',
+                'Jacob', 'Thomas', 'Emily', 'Lily', 'Ava', 'Isla', 'Alfie', 'Olivia', 'Jessica',
+                'Riley', 'William', 'James', 'Geoffrey', 'Lisa', 'Benjamin', 'Stacey', 'Lucy'
+            ]
+
+            last_names = [
+                'Brown', 'Smith', 'Patel', 'Jones', 'Williams', 'Johnson', 'Taylor', 'Thomas',
+                'Roberts', 'Khan', 'Lewis', 'Jackson', 'Clarke', 'James', 'Phillips', 'Wilson',
+                'Ali', 'Mason', 'Mitchell', 'Rose', 'Davis', 'Davies', 'Rodriguez', 'Cox', 'Alexander'
+            ]
+
+            for i in range(len(first_names)):
+                tmp_email = first_names[i].lower() + "." + last_names[i].lower() + "@example.com"
+                # tmp_pass = ''.join(random.choice(string.ascii_lowercase + string.digits) for i in range(10))
+                tmp_pass = first_names[i]
+                user = User(
+                    first_name = first_names[i],
+                    last_name = last_names[i],
+                    email = tmp_email,
+                    password_hash = generate_password_hash(tmp_pass),
+                    roles = [user_role, ]
+                )
+                db.session.add(user)            
+            db.session.commit()
     return
